+++ src/cic_grid.cpp
@@ -7,14 +7,14 @@
 void cic_grid::populate_delta_grid() {
     real_grid.reset_zero();
 
-    for (int p = 0; p < plist->num_particles; ++p) {
-        int i = modulo(plist->x->index(p, 0), n);
-        int j = modulo(plist->x->index(p, 1), n);
-        int k = modulo(plist->x->index(p, 2), n);
+    for (long p = 0; p < plist->num_particles; ++p) {
+        long i = modulo(plist->x->index(p, 0), n);
+        long j = modulo(plist->x->index(p, 1), n);
+        long k = modulo(plist->x->index(p, 2), n);
 
-        int i_neighbor = (i + 1) % n;
-        int j_neighbor = (j + 1) % n;
-        int k_neighbor = (k + 1) % n;
+        long i_neighbor = (i + 1) % n;
+        long j_neighbor = (j + 1) % n;
+        long k_neighbor = (k + 1) % n;
 
         double dx = plist->x->index(p, 0) - i;
         double dy = plist->x->index(p, 1) - j;
@@ -38,9 +38,9 @@ void cic_grid::populate_delta_grid() {
     }
 
 #pragma omp parallel for
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
-            for (int k = 0; k < n; ++k) {
+    for (long i = 0; i < n; ++i) {
+        for (long j = 0; j < n; ++j) {
+            for (long k = 0; k < n; ++k) {
                 real_grid(i, j, k) = real_grid(i, j, k) / avg_particle_density - 1;
             }
         }
