+++ src/grid.cpp
@@ -17,9 +17,9 @@ void grid::ifft() {
     double norm = n * n * n;
 
 #pragma omp parallel for
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
-            for (int k = 0; k < n; ++k) {
+    for (long i = 0; i < n; ++i) {
+        for (long j = 0; j < n; ++j) {
+            for (long k = 0; k < n; ++k) {
                 real_grid(i, j, k) /= norm;
             }
         }
@@ -30,9 +30,9 @@ void grid::apply_greens_func(double a) {
     double factor = -3 * cosmo.Omega_0 / (8 * a);
 
 #pragma omp parallel for
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
-            for (int k = 0; k < (n / 2 + 1); ++k) {
+    for (long i = 0; i < n; ++i) {
+        for (long j = 0; j < n; ++j) {
+            for (long k = 0; k < (n / 2 + 1); ++k) {
                 double sin_x = sin(M_PI * i / n);
                 double sin_y = sin(M_PI * j / n);
                 double sin_z = sin(M_PI * k / n);
@@ -61,7 +61,7 @@ void grid::compute_potential(double a) {
 
 void grid::rebound_positions() {
 #pragma omp parallel for
-    for (int i = 0; i < plist->num_particles; ++i) {
+    for (long i = 0; i < plist->num_particles; ++i) {
         plist->x->index(i,0) = modulo(plist->x->index(i,0), n);
         plist->x->index(i,1) = modulo(plist->x->index(i,1), n);
         plist->x->index(i,2) = modulo(plist->x->index(i,2), n);
